+++ protocol/lookupkey.go
@@ -63,6 +63,9 @@ func (lk *LookupKey) UnmarshalJSON(data []byte) error {
 //   if err != nil { panic(err) }
 //   # sampleKey.Id => "55e8182ec4413d51676d1ba7480708a48c5b50f4a86b3afb9be6c43c648b373d"
 func (lk *LookupKey) Expand(s3Svc s3iface.S3API, s3Bucket string, s3Prefix string) error {
+	// Expand short key types first, even though the data we get back from AWS
+	// \should\ be expanded already. It just makes searching easier
+	lk.Type = lk.Type.Expand()
 	fullPrefix := fmt.Sprintf("%s%s", s3Prefix, lk)
 	objs, err := s3Svc.ListObjectsV2(&s3.ListObjectsV2Input{
 		Bucket: aws.String(s3Bucket),
@@ -87,8 +90,7 @@ func (lk *LookupKey) Expand(s3Svc s3iface.S3API, s3Bucket string, s3Prefix strin
 // parseRawLookupKey takes a string in the same format that `String()` provides
 // and returns the two sub-components of the Key
 //
-// Expansion does NOT happen here. See `lk.Expand()`
-// if you wish to resolve a partial key
+// Returns an error if the key is improperly formatted
 func parseRawLookupKey(rawKey string) (string, CertType, error) {
 	var (
 		typeInd = 0
@@ -99,6 +101,20 @@ func parseRawLookupKey(rawKey string) (string, CertType, error) {
 		return "", "", fmt.Errorf("unable to parse raw key '%s'", rawKey)
 	}
 	return parts[idInd], CertType(parts[typeInd]), nil
+}
+
+// ParseLookupKey takes a string in the same format that `String()` provides
+// and returns a pointer to a new LookupKey object
+//
+// returns an error if the key is improperly formatted
+//
+// Expansion does NOT happen here. See `lk.Expand()`
+// if you wish to resolve a partial key
+func ParseLookupKey(rawKey string) (*LookupKey, error) {
+	var err error
+	lk := &LookupKey{}
+	lk.Id, lk.Type, err = parseRawLookupKey(rawKey)
+	return lk, err
 }
 
 // GenerateLookupKey creates a 64-character long sha256sum key